+++ src/main/kotlin/com/github/jcornaz/kable/impl/BiKeyMap.kt
@@ -17,9 +17,10 @@
  * along with Kable.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package kable
+package com.github.jcornaz.kable.impl
 
-import kable.Table.Entry
+import com.github.jcornaz.kable.Table.Entry
+import com.github.jcornaz.kable.util.entry
 
 /**
  * Implementation of a [Table] backed with [Map] where keys are row-column pairs