+++ src/test/kotlin/com/github/jcornaz/kable/impl/EmptyTableTest.kt
@@ -1,5 +1,9 @@
-package kable
+package com.github.jcornaz.kable.impl
 
+import com.github.jcornaz.kable.util.emptyTable
+import com.github.jcornaz.kable.util.entry
+import com.github.jcornaz.kable.util.minus
+import com.github.jcornaz.kable.util.tableOf
 import org.junit.Assert.*
 import org.junit.Test