+++ src/test/kotlin/com/github/jcornaz/kable/impl/ThreeMapTableTest.kt
@@ -1,5 +1,9 @@
-package kable
+package com.github.jcornaz.kable.impl
 
+import com.github.jcornaz.kable.Table
+import com.github.jcornaz.kable.util.TableTest
+import com.github.jcornaz.kable.util.entry
+import com.github.jcornaz.kable.util.tableOf
 import org.junit.Assert
 import org.junit.Test