+++ src/test/kotlin/com/github/jcornaz/kable/util/TablesTest.kt
@@ -1,5 +1,6 @@
-package kable
+package com.github.jcornaz.kable.util
 
+import com.github.jcornaz.kable.Table
 import org.junit.Assert.*
 import org.junit.Test
 import java.util.*