+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change log
-## 1.2-SNAPSHOT (unreleased)
-### Added
+## 2.0-SNAPSHOT (unreleased)
+### Add
 * Get operator for rows : `table[row]`. It also makes possible to retrieve a value with `table[row][col]` which is equivalent to `table[row, col]`.
 * Mutable tables
+
+### Breaking changes
+* Packages renamed 
+    * Root api : `com.github.jcornaz.kable` (instead of `kable`)
+    * Implementations class : `com.github.jcornaz.kable.impl` (instead of `kable`)
+    * Utilities functions : `com.github.jcornaz.kable.util` (instead of `kable`)
 
 ## 1.1.0 (2017-03-31)
 ### Added